Interpreter

Description

This module contains all of the necissary functions and constants for evaluating the AST created by the Parser.parseAll function.

Documentation

type Env = Map String ([YodaVal] -> YodaVal, Int)

A type alias for a Haskell Map. The keys are strings, the value is a tuple containing the function, and its arity.

defaultEnv :: Env

Yoda built-in functions, or standard library.

yvalEqual :: [YodaVal] -> YodaVal

The equality function for Yoda.

numericBinop :: (Int -> Int -> Int) -> [YodaVal] -> YodaVal

Converts a Haskell function to a Yoda-executable function.

execute :: [YodaVal] -> String -> ([YodaVal] -> YodaVal, Int) -> [YodaVal]

Executes a function

languageFunc :: YodaVal -> Env -> [YodaVal] -> YodaVal

Converts a Yoda lambda to a function that can be named and added to the defaultEnv.

evalExpr :: Env -> YodaVal -> [YodaVal] -> ([YodaVal], Env)

Evaluates a single Yoda expression, with the environment and the stack.

run :: [YodaVal] -> [YodaVal] -> Env -> ([YodaVal], Env)

Evaluates multiple Yoda expressions.

eval :: String -> [YodaVal] -> Env -> ([YodaVal], Env)

Evalutates a string.