Parser

Description

This module contains all of the necissary logic for parsing Yoda, turning it into an AST, and attaching some metadata. I use Parsec to parse Yoda.

Documentation

symbol :: Parser Char

Matches one of the symbols that are allowed in Yoda identifiers.

parseId :: Parser YodaVal

Matches a Yoda identifier: The first character is a letter or a symbol, the rest of the characters are either a letter, a digit, or a symbol.

parseNumber :: Parser YodaVal

Parsers a integer or a floating point number. Minor oddity: A negitive sign is an underscore in Yoda: _1 is a negative number.

spaces :: Parser ()

Matches and ignores whitespace.

parseString :: Parser YodaVal

Parses a Yoda string. Only double quotes work.

parseList :: Parser YodaVal

Parses a bunch of space separated expressions, starting and ending with a space.

parseExpr :: Parser YodaVal

Parses either a number, an identifier, or a block.

parseExprs :: Parser [YodaVal]

Parses multiple expressions.

parseAll :: String -> [YodaVal]

Parses a string, using the PArsec parse function.