Structures

Description

The data type represented here shows the possible syntactic forms of Yoda, represented in Haskell for exectution and storage.

Documentation

data YodaVal

Represents the possible types of data in Yoda

Constructors

Number Int
Represents a number in Yoda.

Id String
Represents an identifier in Yoda.

Func [YodaVal]
A function in Yoda is a list of YodaVals, in this case the parsed code inside the function. This code can be directly fed into the interpreter to be run.

Str String
A Yoda String.

Decimal Float
This represents a decimal in Yoda.

Error String
An error type, for error correction in Yoda.

unpackNumber :: YodaVal -> Int

unpackNumber x returns the unwrapped Number x, or 0.

unpackString :: YodaVal -> String

unpackString x changes x into a string if it is not, using show. Otherwise if it is a Str, Error, or Id, returns the value.

unpackDecimal :: YodaVal -> Float

unpackDecimal x makes a Number into a Float, or just returns the unwrapped value of a Decimal. If x is neither a Decimal or a Number, it returns 0.0.

getBody :: YodaVal -> [YodaVal]

Returns a function body, otherwise throws an error

showVal :: YodaVal -> String

Formats a Yoda value.